moonjyoung 230130 ver.01

+++ src/main/java/com/deliverygig/moonjyoung/vo/store/ShowStoreListVO.java
@@ -11,7 +11,7 @@ public class ShowStoreListVO {
     private Long storeSeq;
     private String storeName;
     private LocalTime storeCloseTime;
-    private Double discount;
+    private Integer discount;
     private Integer status;
     // private List<StoreClosedDayInfoVO> storeClosedDay ; 
     private Long simgSiSeq;
